Renamed NVTX to GPU_TX_MARKERS and added note for AMD GPUs
