+++ dace/codegen/instrumentation/__init__.py
@@ -7,6 +7,6 @@ from .likwid import LIKWIDInstrumentationCPU, LIKWIDInstrumentationGPU
 from .timer import TimerProvider
 from .gpu_events import GPUEventProvider
 from .fpga import FPGAInstrumentationProvider
-from .nvtx import NVTXProvider
+from .gpu_tx_markers import GPUTXMarkersProvider
 
 from .data.data_dump import SaveProvider, RestoreProvider
