+++ dace/codegen/instrumentation/gpu_events.py
@@ -3,19 +3,14 @@ from typing import Union
 from dace import config, dtypes, registry
 from dace.codegen.prettycode import CodeIOStream
 from dace.sdfg import nodes, is_devicelevel_gpu
-from dace.codegen import common
-from dace.codegen.instrumentation.nvtx import NVTXProvider
+from dace.codegen.instrumentation.gpu_tx_markers import GPUTXMarkersProvider
 from dace.sdfg.sdfg import SDFG
 from dace.sdfg.state import ControlFlowRegion, SDFGState
 
 
 @registry.autoregister_params(type=dtypes.InstrumentationType.GPU_Events)
-class GPUEventProvider(NVTXProvider):
+class GPUEventProvider(GPUTXMarkersProvider):
     """ Timing instrumentation that reports GPU/copy time using CUDA/HIP events. """
-
-    def __init__(self):
-        self.backend = common.get_gpu_backend()
-        super().__init__()
 
     def on_sdfg_begin(self, sdfg: SDFG, local_stream: CodeIOStream, global_stream: CodeIOStream, codegen) -> None:
         if self.backend == 'cuda':
