+++ dace/dtypes.py
@@ -165,7 +165,7 @@ class InstrumentationType(aenum.AutoNumberEnum):
     LIKWID_GPU = ()
     GPU_Events = ()
     FPGA = ()
-    NVTX = ()
+    GPU_TX_MARKERS = ()
 
 
 @undefined_safe_enum
